requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyGithub==2.1.1
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
